refactor: separate tooltip
