+++ assets/i18n/en.i18n.json
@@ -333,6 +333,7 @@
       }
     },
     "accounting": {
+      "tooltip": "Have you discussed the departure time with the party members?\nClick this button to request accounting.\nAccounting requests are only available to the payer.\nPlease be careful not to make a wrong request.",
       "before_confirm": "You cannot request accounting before confirming the departure time",
       "dutch": {
         "title": "Dutch Pay",
