+++ assets/i18n/ko.i18n.json
@@ -330,6 +330,7 @@
       }
     },
     "accounting": {
+      "tooltip": "파티원들과의 동승을 잘 마치셨나요?\n이 버튼을 눌러 정산을 요청하세요.\n정산 요청은 결제자만 이용 해야 해요.\n잘못된 요청을 하지 않도록 신중하세요.",
       "before_confirm": "출발 시간을 확정하지 않아 정산을 요청할 수 없습니다",
       "dutch": {
         "title": "정산하기",
