+++ lib/app/modules/chat/presentation/widgets/accounting_button.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:pot_g/app/modules/chat/domain/entities/pot_info_entity.dart';
+import 'package:pot_g/app/modules/chat/presentation/widgets/tooltip_overlay.dart';
 import 'package:pot_g/app/modules/common/presentation/extensions/toast.dart';
 import 'package:pot_g/app/modules/common/presentation/widgets/pot_icon_button.dart';
 import 'package:pot_g/app/router.gr.dart';
@@ -8,7 +9,7 @@ import 'package:pot_g/app/values/palette.dart';
 import 'package:pot_g/gen/assets.gen.dart';
 import 'package:pot_g/gen/strings.g.dart';
 
-class AccountingButton extends StatelessWidget {
+class AccountingButton extends StatefulWidget {
   const AccountingButton({super.key, required this.pot});
 
   final PotInfoEntity pot;
@@ -35,14 +36,25 @@ class AccountingButton extends StatelessWidget {
   }
 
   @override
+  State<AccountingButton> createState() => _AccountingButtonState();
+}
+
+class _AccountingButtonState extends State<AccountingButton> {
+  final _controller = OverlayPortalController();
+
+  @override
   Widget build(BuildContext context) {
-    return PotIconButton(
-      icon: Assets.icons.dollar.svg(
-        colorFilter: ColorFilter.mode(Palette.grey, BlendMode.srcIn),
+    return TooltipOverlay(
+      controller: _controller,
+      content: Text(context.t.chat_room.accounting.tooltip),
+      child: PotIconButton(
+        icon: Assets.icons.dollar.svg(
+          colorFilter: ColorFilter.mode(Palette.grey, BlendMode.srcIn),
+        ),
+        onPressed: () async {
+          await AccountingButton.setAccounting(context, widget.pot);
+        },
       ),
-      onPressed: () async {
-        await setAccounting(context, pot);
-      },
     );
   }
 }
